refactor(chain): share finalized RPC transport lifecycle
